services/integration/pyproject.toml
[project]
name = "integration-service"
version = "1.0.0"
description = "Cross-App Integration Layer for PhotoIdentifier Platform"
dependencies = [
    "aiokafka>=0.8.11",
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "jsonschema>=4.20.0",
    "avro-python3>=1.10.2",
    "confluent-kafka>=2.3.0",
    "httpx>=0.25.0",
    "python-jose[cryptography]>=3.3.0",
    "sqlalchemy>=2.0.23",
    "asyncpg>=0.29.0",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
